Only writes the file if it doesn't already exist

+++ src/main/java/com/github/brotherlogic/pictureframe/DropboxConnector.java
@@ -17,8 +17,7 @@ public class DropboxConnector {
 
 	DbxClient client;
 
-	DbxRequestConfig config = new DbxRequestConfig("TuckerPictureFrame/1.0",
-			Locale.getDefault().toString());
+	DbxRequestConfig config = new DbxRequestConfig("TuckerPictureFrame/1.0", Locale.getDefault().toString());
 
 	public DropboxConnector(String token) {
 		client = new DbxClient(config, token);
@@ -31,11 +30,9 @@ public class DropboxConnector {
 		// Have the user sign in and authorize your app.
 		String authorizeUrl = webAuth.start();
 		System.out.println("1. Go to: " + authorizeUrl);
-		System.out
-				.println("2. Click \"Allow\" (you might have to log in first)");
+		System.out.println("2. Click \"Allow\" (you might have to log in first)");
 		System.out.println("3. Copy the authorization code.");
-		String code = new BufferedReader(new InputStreamReader(System.in))
-				.readLine().trim();
+		String code = new BufferedReader(new InputStreamReader(System.in)).readLine().trim();
 
 		// This will fail if the user enters an invalid authorization code.
 		DbxAuthFinish authFinish = webAuth.finish(code);
@@ -43,17 +40,18 @@ public class DropboxConnector {
 		System.out.println("TOKEN = " + accessToken);
 	}
 
-	public void syncFolder(String inputDirectory, File outputDirectory)
-			throws Exception {
-		DbxEntry.WithChildren listing = client
-				.getMetadataWithChildren(inputDirectory);
+	public void syncFolder(String inputDirectory, File outputDirectory) throws Exception {
+		DbxEntry.WithChildren listing = client.getMetadataWithChildren(inputDirectory);
 		for (DbxEntry child : listing.children) {
-			FileOutputStream outputStream = new FileOutputStream(
-					outputDirectory.getAbsolutePath() + "/" + child.name);
-			try {
-				client.getFile(child.path, null, outputStream);
-			} finally {
-				outputStream.close();
+			File f = new File(outputDirectory.getAbsolutePath() + "/" + child.name);
+			if (!f.exists()) {
+				FileOutputStream outputStream = new FileOutputStream(
+						outputDirectory.getAbsolutePath() + "/" + child.name);
+				try {
+					client.getFile(child.path, null, outputStream);
+				} finally {
+					outputStream.close();
+				}
 			}
 		}
 	}
